requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
crispy-bootstrap4==2025.6
Django==5.2.5
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-debug-panel==0.8.3
django-debug-toolbar==6.0.0
django-ipware==7.0.1
django-rosetta==0.10.2
django-silk==5.4.1
djangorestframework==3.16.1
envparse==0.2.0
fontawesomefree==6.6.0
gprof2dot==2025.4.14
idna==3.10
pillow==11.3.0
polib==1.2.0
PyPDF2==3.0.1
python-dotenv==1.1.1
python-ipware==3.0.0
reportlab==4.4.3
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
